Remove call to acquire_credentials.
Remove the "gss.acquire_credentials" call, which a) should never be needed, and b) will sometimes fail.

This call was failing in my environment, and to the best of my knowledge, even if it succeeds, the credentials it acquires will never be used. The default usage for acquire_credentials, if none is specified, is "acceptor". However, acceptor credentials at the GSS layer are never used by init_sec_context.