Always forward Kerberos tickets.
Adding ":delegate => true" to the init_context call causes Kerberos tickets to be forwarded to the remote host.  This may or may not be a good idea.

Pro: This is consistent with the behavior of the gem under ruby1.8.

Con: This might be a security issue, and tickets should really only be forwarded if the user specifies that this is what they want.

It wasn't immediately obvious to me how to make this optional, so I'm submitting this the "dumb" version of this change with the above caveats.